# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "manypost API",
    "version": "0.0.1",
    "description": "API do manypost — agendamento/publicação multicanal. Humanos usam sessão Clerk; máquinas usam `Authorization: Bearer mp_live_…`. Erros seguem problem+json (RFC 9457): o campo `title` carrega o código estável do erro."
  },
  "servers": [
    {
      "url": "/",
      "description": "mesma origem (proxy do Next)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "sessão Clerk humana — Authorization: Bearer <token>"
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key mp_live_ para superfícies de máquina — Authorization: Bearer <chave>"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "description": "código estável do erro",
            "example": "common.not_found"
          },
          "status": {
            "type": "integer",
            "example": 404
          },
          "detail": {
            "type": "string",
            "example": "post não encontrado"
          },
          "extra": {
            "description": "campos extras (ex.: issues de validação)"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "FeedItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "channelId": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "example": "PUBLISHED"
          },
          "publishAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "publishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "quando a entrega aconteceu"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "última mutação da linha — ordena atividade recente"
          },
          "text": {
            "type": "string"
          },
          "mediaCount": {
            "type": "integer"
          },
          "mediaPreview": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "image",
                  "video"
                ]
              },
              "url": {
                "type": "string"
              },
              "mime": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "alt": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "type",
              "url",
              "mime",
              "alt"
            ],
            "description": "primeira mídia do conteúdo para reconhecimento visual do card"
          },
          "externalId": {
            "type": [
              "string",
              "null"
            ]
          },
          "releaseUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "errorClass": {
            "type": [
              "string",
              "null"
            ]
          },
          "errorMessage": {
            "type": [
              "string",
              "null"
            ]
          },
          "attemptCount": {
            "type": "integer"
          },
          "group": {
            "type": "object",
            "properties": {
              "state": {
                "type": "string"
              },
              "origin": {
                "type": "string",
                "example": "WEB"
              },
              "awaitingApproval": {
                "type": "boolean"
              }
            },
            "required": [
              "state",
              "origin",
              "awaitingApproval"
            ]
          },
          "channel": {
            "type": "object",
            "properties": {
              "provider": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "username": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "avatarUrl": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "provider",
              "name",
              "username",
              "avatarUrl"
            ]
          }
        },
        "required": [
          "id",
          "groupId",
          "channelId",
          "state",
          "publishAt",
          "publishedAt",
          "updatedAt",
          "text",
          "mediaCount",
          "mediaPreview",
          "externalId",
          "releaseUrl",
          "errorClass",
          "errorMessage",
          "attemptCount",
          "group",
          "channel"
        ]
      },
      "PublicationFeed": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedItem"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "cursor keyset da próxima página"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
...
      "PubFeedItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "channelId": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "example": "PUBLISHED"
          },
          "publishAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "publishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "quando a entrega de fato aconteceu (null antes de publicar)"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "última mutação da linha — ordena atividade recente"
          },
          "text": {
            "type": "string"
          },
          "mediaCount": {
            "type": "integer"
          },
          "mediaPreview": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "image",
                  "video"
                ]
              },
              "url": {
                "type": "string"
              },
              "mime": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "alt": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "type",
              "url",
              "mime",
              "alt"
            ],
            "description": "primeira mídia do conteúdo, quando existe"
          },
          "externalId": {
            "type": [
              "string",
              "null"
            ]
          },
          "releaseUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "errorClass": {
            "type": [
              "string",
              "null"
            ]
          },
          "errorMessage": {
            "type": [
              "string",
              "null"
            ]
          },
          "attemptCount": {
            "type": "integer"
          },
          "group": {
            "type": "object",
            "properties": {
              "state": {
                "type": "string"
              },
              "origin": {
                "type": "string"
              },
              "awaitingApproval": {
                "type": "boolean"
              }
            },
            "required": [
              "state",
              "origin",
              "awaitingApproval"
            ]
          },
          "channel": {
            "type": "object",
            "properties": {
              "provider": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "username": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "avatarUrl": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "provider",
              "name",
              "username",
              "avatarUrl"
            ]
          }
        },
        "required": [
          "id",
          "groupId",
          "channelId",
          "state",
          "publishAt",
          "publishedAt",
          "updatedAt",
          "text",
          "mediaCount",
          "mediaPreview",
          "externalId",
          "releaseUrl",
          "errorClass",
          "errorMessage",
          "attemptCount",
          "group",
          "channel"
        ]
      },
      "PubPublicationFeed": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PubFeedItem"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/publications": {
      "get": {
        "tags": [
          "publications"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Feed flat de publicações (calendário/kanban)",
        "description": "Uma linha por publicação (o cliente agrupa por groupId). Cada item embute grupo (state/origin/awaitingApproval) e canal. Paginação keyset por (publishAt, id).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "início do período (ISO 8601)"
            },
            "required": false,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "fim do período (ISO 8601)"
            },
            "required": false,
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "csv de estados de publicação",
              "example": "SCHEDULED,PUBLISHED"
            },
            "required": false,
            "name": "state",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "csv de ids de canal"
            },
            "required": false,
            "name": "channelId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "cursor keyset da página anterior"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "1–200 (default 100)",
              "example": "100"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "página do feed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationFeed"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/v1/publications": {
      "get": {
        "tags": [
          "public-publications"
        ],
        "summary": "Feed flat de publicações — status por canal (escopo posts:read)",
        "description": "Uma linha por publicação; paginação keyset por (publishAt, id).",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "SCHEDULED,PUBLISHED"
            },
            "required": false,
            "name": "state",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "channelId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "100"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "página do feed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PubPublicationFeed"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "papel insuficiente (requer ADMIN/OWNER)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "rate limit — aguarde e tente de novo",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}